Failure Handling Rework (Discord autolaunch)
Alright so a lot of stuff had to be done in order to get this to work.

So what is now happening here is that the code itself has literally changed it's way to work. Blame this on the discord-rpc package, which has no way to handle a connection failure properly (e.g. all reconnecting attempts result in failure), and therefore the only way to free the port that was occupied by the client is to actually restart the program.

Due to this another dependency was added.

BEFORE: All the stuff in the "master process" - the process that is launched when we run 'npm start'. If RPC errors, we need to close and reopen.

NOW: When invoked, the "master process" deploys a slave process (with the environment variable named "terminal" set to 0). That process attempts a first connection. In case of failure and if set up to autolaunch discord, it exits with a status code of 1, which leads the master process to launch another slave process (this time with the env variable set to "1"). A second attempt is then made, which in case of failure leads to the program to give up and display a notification to the user.

 OTHER STUFF:

- This stuff only works in a limited number of installations, so the configurations are set to false by default.

- LOGGING: Instead of erasing the file every time roPresence is launched, it will append an header instead.